This fixes a broken commit and fixes scripts/before_i_commit.sh
Normally, I would not put two fixes into one commit - however, git master
was broken. The fix in txscapy.py fixes it such that I can run tests at all.
The fix in scripts/before_i_commit.sh allows me to actually run the thing.
It now depends on bash and as a result, it works properly. If we actually
want this this to run on an older unix machine, such a user can fix it to work
with a /bin/sh that isn't based on bash or later implementations.

The previous version of scripts/before_i_commit.sh erased all .yamloo files
and as a result, erased a lot of import data. If we run tests, we should not
ever erase the data automatically.