This is a possible fix for https://trac.torproject.org/projects/tor/ticket/7664
In my tests, running a deck with this patch gave the desired output (one report
by test). This doesn't mean this fix is the best one, and I don't feel
confortable enough with the codebase to even know if this is the right
direction.

The underlying problem seems to be that this code is using a global variable
"reports" (imported from config.py) to store the report filename, and this is
getting overwritten for each test that's added. Because of that, the last
reportfile described in the deck (for before_i_commit.deck it was
http_host.yamloo) was beeing used every time. Since the reporter saved a backup
of the file if it already existed, it saved the output of the last test (for
before_i_commit.deck it was dns_tamper) in and .old file.