Fixes reportfile creation when the reportfile name is not defined
My last commit (2becc29d40f8) broke the reporting when the filename wasn't
defined. I took the filename creation logic from
ooni/config.py@generateReportFilenames to use when the command line (or test
deck) does not define a name for the report file.

If this is the right way to go, this code probably needs a cleanup, at least to
avoid duplication of this code.