Get test version from net_test_file and parse all options
Adds helper method _parseNetTestOptions to NetTestLoader in order
to assemble the various options and flags into a single object

Also add functionality to read the test name and version from the
NetTestCase

Check that the required option is not None

localOptions is a dict, and the required option may be None.
This adds a test for this case.