Update user agent strings for user agent randomization
They are now taken from the PETS2010 paper "How unique is your browser" as the
user agents with the largest anonymity set.
Note: this is probably no longer true as of today, but I could not find a
better source for mostly used user agent strings and it's for sure better than
what we had before (We were considering highly used Opera 8.1).