Wait until Tor has started before adding reporters
As we get the client IP from Tor, we need to wait until Tor has
bootstrapped (if start_tor is set) before setting up the reporters.

Note that we should try and get the client IP by an alternate
mechanism if start_tor is not enabled in ooniprobe.conf or if we
are unable to get the client IP from Tor.