Drops the hierarchy by reportFormatVersion
See also: https://github.com/TheTorProject/ooni-backend/issues/20

This was not implemented in ooni-probe or ooni-backend, and is
redundant to the report, though if we decide to change the report
header in the future we may need to handle this case in parsers.

Adding new fields should generally not be a problem, though
ooni-backend presently expects a fixed length report header.