Add postProcessor support to NetTest
If a NetTestCase has multiple test_ methods, then the report write shall
not occur until all of the individual Measurements have completed.

This also changes the postProcessor API slightly; postProcessors are now
expected to return a report instance. If a postProcessor is not
supplied, then the report object from the instanced NetTestCase class is
passed to Report.write()