Add Tests for Lantern (https://getlantern.org)
LanternBootstrapProcessDirector monitors Lanterns output (stdout, stderr) and
calls back when Lantern has bootstrapped. If a timeout is configured and
Lantern has not bootstrapped before it fires, then the process director will
errback instead.

LanternTest has a single test test_lantern_circumvent that waits for 20 seconds
for LanternBootstrapProcessDirector to call back when Lantern has bootstrapped,
and then makes a single request to http://google.com

The report contains the output of stdout and stderr, if Lantern bootstrapped
successfully, the page body requested, or a failure and reason.