Do not reset timer when output is received
If the timer is reset upon output received, eventually OONI's
built-in timer will fire and consider the the test a failure.

As demonstrated in the example code, the process is launched by
a Measurement method. Long-running processes probably shouldn't
be started inside Measurements, and it might make more sense to
launch processes as part of the NetTest setUp method instead.