OpenVPN connection circumvention test
This test attempts to connect to an OpenVPN server specified in an
OpenVPN config file. A HTTP request is sent if the VPN connection is
successful.

This basic test can be expanded to parse more of the output from the
OpenVPN client. The output could provide more information about why
a connection might be failing.