Really handleAllFailures()
handleAllFailures() only handles failures which are explicitly named. Maybe it
should be renamed or emit a log message when an unknown failure is reraised
from `failure.trap()`

The known failure list has been converted to a list of tuples as
the order of the failure checking is important. Some Failures subclass
other failures and so they should be listed from most specific failure
to the least specific.