psiphon: use the python provided by sys.executable
There may be more than one Python (for example on MacOS I have both
the system Python and the one installed with brew).

Make sure we use the same Python with which ooni-probe was started
such that modules installed with pip are found.

Updated to avoid setting the temporary file executable and to avoid
removing my Python interpreter :).