psiphon: fix bug that psiphon hangs ooniprobe -ng
Marked my fix with XXX because I'm not sure whether this is
the proper approach, or whether it would be better to work on
a patch to teach ooniprobe's core to deal with that case.

For now, and given that a psiphon MVP must be delivered as
soon as possible, this fix will do.