Make inputs a shared iterator across all NetTestCase measurements
This commit makes the inputs iterator shared by all the measurement
tests within a NetTestCase. Per issue 503, In order to make bisection
style testing functional the individual tests need to be able to pass
data to the main test_case.inputs generator that seeds each
measurement. This provides opportunities for each measurement test in
a NetTestCase to communicate with the iterator to do things such as
adding additional values to be passed to measurement tests as a later
input.