Add support for loglevels.
Small refactor of the logging facilities

Note: now all messages printed with print() are not output to stdout by default
and all messages that are not printed via the ooni logging infrastructure that
are not critical are also ignored.

This is useful to avoid cluttering the CLI with too many messages and also to
reduce the size of the logs.

This implements: #502