Smear load generated by scheduler
* Smear load generated by scheduler, see #630

It runs ASAP after the installation. It survives reboot in predictable
way: avoids double-run and does not skip scheduler cycle on restart.
It supports non-24/7 operations as it targets running @daily task within
2.4 hours window, so the probe that was down will likely run a task
on boot and boot time is random enough for load smearing purposes.

* Schedule decks right after initialisation, see #630

That's required as refresh_deck_list() does not schedule anything if the
`config` is not `is_initialized()`.

Another possible way to implement this behavior is to postpone
RefreshDeckList.task run checking config.is_initialized() in
should_run().