Should close #639.
This is my first pull request to a major project, please be mindful of
that. If web_connectivity.py was issued a URL that urlparse() cannot
disseminate into individual elements, an exception will be raised.
Furthermore, if the supplied URL does not start with 'http', InvalidURL
will be raised with a message addressing this particular issue.